Add files via upload
Nesta atualização algumas linhas de código do index foram separadas em outros arquivos, e também foi iniciado um banco de dados
