+++ package.json
@@ -11,7 +11,9 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "express": "^4.18.2"
+    "@prisma/client": "^4.14.1",
+    "express": "^4.18.2",
+    "prisma": "^4.14.1"
   },
   "devDependencies": {
     "nodemon": "^2.0.22"
